% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorTestPvalue.R
\name{CorTestPvalue}
\alias{CorTestPvalue}
\title{p-value of correlation test}
\usage{
CorTestPvalue(data, depenVaColname, indepenVaColname)
}
\arguments{
\item{data}{Your dataset}

\item{depenVaColname}{The varibale that we are interested in}

\item{indepenVaColname}{The varibale that we want to know the correlation p-vlaue with the variable that we are interested in.}
}
\value{
a list
}
\description{
This function is designed to get p-value by doing cor.test to a variable with the varibale that we are interested in.
}
\examples{
CorTestPvalue(exampledata,"Weight (in kg)","Olive oil (in g)")


}
