% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{Personal Health Data Sheet}
\format{A data frame with 15 rows and 15 variables}
\source{
\url{https://docs.google.com/spreadsheets/d/1S1TSS8Wwx8fU1C0-S30zGzZXj_R8zSM-BQSzFNZzGic/edit#gid=0}
}
\description{
This dataset describes daily health data related to my weight
}
\details{
\itemize{
  \item my weight(in kg) from 09/13/2017 to 09/27/2017
  \item Olive oil(in g) intake from 09/13/2017 to 09/27/2017
  \item Red meat(in g) intake from 09/13/2017 to 09/27/2017
  \item Other Protein(in g) intake from 09/13/2017 to 09/27/2017
  \item Protein (in g) intake from 09/13/2017 to 09/27/2017
  \item Vegetables (in g) intake from 09/13/2017 to 09/27/2017
  \item Rice and Grain (in g) intake from 09/13/2017 to 09/27/2017
  \item Water (in cup) intake from 09/13/2017 to 09/27/2017
  \item Snacks (in calories) intake from 09/13/2017 to 09/27/2017
  \item Fitness(in mins) time from 09/13/2017 to 09/27/2017
  \item Study Hours(in hrs) time from 09/13/2017 to 09/27/2017
  \item Entertament (in hrs) time from 09/13/2017 to 09/27/2017
  \item Sleep hours (in hrs) time from 09/13/2017 to 09/27/2017
  \item Sleep before 11pm or not from 09/13/2017 to 09/27/2017

}
}
\keyword{datasets}
