% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scatter.R
\name{ggplot_scatter}
\alias{ggplot_scatter}
\title{sctter plot on two variables}
\usage{
ggplot_scatter()
}
\value{
graph
}
\description{
sctter plot between fitness time and protein intake and fitness time and rice and grain intake
}
