% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiCorTest.R
\name{multiCorTest}
\alias{multiCorTest}
\title{mutiple correlation test}
\usage{
multiCorTest(data, depenVaColname)
}
\arguments{
\item{data}{Your dataset}

\item{depenVaColname}{The varibale that we are interested in}
}
\description{
This function is used when we are interested in one column in the dataset and want to test its correlations with variables in other columns.
}
\examples{
multiCorTest(mydata,"Weight (in kg)")
}
